requirements.txt
awscli
black
flake8
ipython
isort
jupyterlab
loguru
matplotlib
mkdocs
notebook
numpy
pandas
pip
python-dotenv
scikit-learn
tqdm
typer
joblib
openai
json
catboost
# packet
-e .
